pyproject.toml
[project]
name = "doip_server"
version = "0.1.0"
description = "DoIP Server and Client implementation with YAML configuration"
authors = [
    {name = "Alonso Ramirez",email = "alonsoram07@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "scapy (>=2.6.1,<3.0.0)",
    "PyYAML (>=6.0,<7.0)"
]

[tool.poetry]
packages = [{include = "doip_server", from = "src"}]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-cov = "^6.0.0"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--disable-warnings"
]
markers = [
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "slow: marks tests as slow"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
main = 'doip_server.main:main'
doip_server = 'doip_server.doip_server:start_doip_server'
doip_client = 'doip_server.doip_client:start_doip_client'
demo = 'doip_server.demo:main'
validate_config = 'doip_server.validate_config:main'